overlay for mousefilter

+++ src/RemoveRedFilter.java
@@ -8,7 +8,7 @@ public class RemoveRedFilter implements PixelFilter {
         short[][] blackPixels = new short[img.getHeight()][img.getWidth()];
 
         for(int i = 0; i < blackPixels.length; i++) {
-            Arrays.fill(blackPixels[i], 100);
+            Arrays.fill(blackPixels[i], (short) 100);
         }
 
         img.setRedChannel(blackPixels);
